fix: refine mobile sidebar navigation
