+++ src/components/navigation/NavUser.vue
@@ -1,7 +1,6 @@
 <script setup lang="ts">
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -14,14 +13,6 @@ import {
   DropdownMenuSubTrigger,
 } from "@/components/ui/dropdown-menu";
 import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import {
   SidebarMenu,
   SidebarMenuButton,
   SidebarMenuItem,
@@ -30,9 +21,8 @@ import {
 import { authManager } from "@/plugins/auth";
 import { api } from "@/plugins/api";
 import { copyToClipboard } from "@/helpers/utils";
+import { useAccountSwitcher } from "@/composables/useAccountSwitcher";
 import { useScrobblingStatus } from "@/composables/useScrobblingStatus";
-import type { SavedAccount } from "@/plugins/auth";
-import type { User } from "@/plugins/api/interfaces";
 import { store } from "@/plugins/store";
 import { Activity, Copy, LogOut, Pencil, Server, UserRound } from "@lucide/vue";
 import { computed, ref } from "vue";
@@ -40,12 +30,10 @@ import { useI18n } from "vue-i18n";
 import { toast } from "vue-sonner";
 import { useRouter } from "vue-router";
 import ProfileAvatarEditor from "@/components/profile/ProfileAvatarEditor.vue";
-import { Input } from "@/components/ui/input";
 import {
   accountAccentClass as getAccountAccentClass,
   accountAccentBackgroundClass as getAccountAccentBackgroundClass,
   accountAccentButtonClass as getAccountAccentButtonClass,
-  accountSwitcherAccentClass,
   getConnectionStatusKey,
 } from "./accountMenu";
 
@@ -54,36 +42,16 @@ const { status: scrobblingStatus, providerSummary } = useScrobblingStatus();
 
 const router = useRouter();
 const { isMobile, setOpenMobile } = useSidebar();
+const {
+  accountInitial,
+  accountName,
+  availableAccounts,
+  loadAccounts,
+  openAccountSwitcher: requestAccountSwitcher,
+  selectAccount,
+} = useAccountSwitcher();
 const accountMenuOpen = ref(false);
-const switchAccountsOpen = ref(false);
-const savedAccounts = computed(() => authManager.getSavedAccounts());
-const serverUsers = ref<User[]>([]);
 const savingAvatar = ref(false);
-const loginUser = ref<User | null>(null);
-const loginPassword = ref("");
-const loginLoading = ref(false);
-const loginError = ref("");
-const loginDialogOpen = ref(false);
-interface AccountOption {
-  user?: User;
-  account?: SavedAccount;
-}
-
-const availableAccounts = computed<AccountOption[]>(() => {
-  if (!serverUsers.value.length) {
-    return savedAccounts.value.map((account) => ({
-      user: undefined,
-      account,
-    }));
-  }
-
-  return serverUsers.value.map((user) => ({
-    user,
-    account: savedAccounts.value.find(
-      (account) => account.username === user.username,
-    ),
-  }));
-});
 
 const displayName =
   store.currentUser?.display_name || store.currentUser?.username || "";
@@ -151,78 +119,14 @@ const copyUsername = async () => {
 
 const openAccountSwitcher = () => {
   accountMenuOpen.value = false;
-  switchAccountsOpen.value = true;
-  loadAccounts();
-};
-
-const loadAccounts = async () => {
-  if (!authManager.isAdmin()) {
-    serverUsers.value = [];
-    return;
+  if (isMobile.value) {
+    setOpenMobile(false);
   }
-  try {
-    serverUsers.value = await api.getAllUsers();
-  } catch {
-    // Non-admin users cannot list all accounts; remembered sessions still work.
-  }
+  requestAccountSwitcher();
 };
 
-const addAccount = () => {
-  authManager.logout();
-};
-
-const switchAccount = (account: SavedAccount | undefined) => {
-  if (!account) return;
-  switchAccountsOpen.value = false;
-  authManager.switchAccount(account);
-};
-
-const selectAccount = (item: { user?: User; account?: SavedAccount }) => {
-  if (item.account) {
-    switchAccount(item.account);
-    return;
-  }
-  if (!item.user) return;
-  loginUser.value = item.user;
-  loginPassword.value = "";
-  loginError.value = "";
-  loginDialogOpen.value = true;
-};
-
-const signInToAccount = async () => {
-  if (!loginUser.value || !loginPassword.value || loginLoading.value) return;
-  loginLoading.value = true;
-  loginError.value = "";
-  try {
-    const result = await api.loginWithCredentials(
-      loginUser.value.username,
-      loginPassword.value,
-    );
-    authManager.setToken(result.token);
-    authManager.setCurrentUser(result.user);
-    loginDialogOpen.value = false;
-    switchAccountsOpen.value = false;
-    window.location.reload();
-  } catch (error: unknown) {
-    loginError.value =
-      error instanceof Error ? error.message : t("auth.login_failed");
-  } finally {
-    loginLoading.value = false;
-  }
-};
-
-const accountName = (account: SavedAccount | User) =>
-  ("display_name" in account ? account.display_name : account.displayName) ||
-  account.username;
-
-const accountInitial = (account: SavedAccount | User) =>
-  accountName(account).charAt(0).toUpperCase() || "U";
-
-const accountAccentClass = (account: SavedAccount | User) =>
+const accountAccentClass = (account: { username: string }) =>
   getAccountAccentClass(account.username);
-
-const accountSwitcherClass = (account: SavedAccount | User) =>
-  accountSwitcherAccentClass(account.username);
 </script>
 
 <template>
@@ -250,7 +154,7 @@ const accountSwitcherClass = (account: SavedAccount | User) =>
                 aria-hidden="true"
               ></span>
               <span
-                v-if="scrobblingStatus.configured"
+                v-if="scrobblingStatus.configured && !accountMenuOpen"
                 class="scrobbling-avatar-glow pointer-events-none absolute -inset-1 rounded-full"
                 :class="[
                   currentAccountAccentBackgroundClass,
@@ -301,12 +205,40 @@ const accountSwitcherClass = (account: SavedAccount | User) =>
               ></div>
               <div class="relative px-3 pb-2">
                 <div class="account-profile-avatar absolute -top-8 left-3">
-                  <ProfileAvatarEditor
-                    :model-value="store.currentUser?.avatar_url"
-                    :avatar-class="`size-20 border-4 ${currentAccountAccentClass}`"
-                    :disabled="store.isIngressSession || savingAvatar"
-                    @update:model-value="handleAvatarUpdate"
+                  <span
+                    v-if="scrobblingStatus.configured && accountMenuOpen"
+                    class="scrobbling-avatar-glow scrobbling-avatar-glow--active pointer-events-none absolute -inset-1 rounded-full"
+                    :class="currentAccountAccentBackgroundClass"
+                    aria-hidden="true"
+                  ></span>
+                  <div class="relative z-10">
+                    <ProfileAvatarEditor
+                      :model-value="store.currentUser?.avatar_url"
+                      :avatar-class="`size-20 border-4 ${currentAccountAccentClass}`"
+                      :disabled="store.isIngressSession || savingAvatar"
+                      @update:model-value="handleAvatarUpdate"
+                    />
+                  </div>
+                </div>
+                <div
+                  v-if="scrobblingStatus.configured"
+                  class="absolute top-2 right-3 left-26 flex min-w-0 items-center gap-1.5 rounded-lg bg-muted/50 px-2 py-1.5"
+                  role="status"
+                  :aria-label="scrobblingLabel"
+                  :title="scrobblingLabel"
+                >
+                  <Activity
+                    class="size-3.5 shrink-0 text-primary"
+                    aria-hidden="true"
                   />
+                  <div class="grid min-w-0 gap-0.5 text-left leading-tight">
+                    <span class="text-[10px] font-medium text-foreground">{{
+                      $t("auth.scrobbling")
+                    }}</span>
+                    <span class="truncate text-[10px] text-muted-foreground">{{
+                      providerSummary
+                    }}</span>
+                  </div>
                 </div>
                 <div class="grid min-w-0 pt-17 text-left leading-tight">
                   <span
@@ -349,22 +281,6 @@ const accountSwitcherClass = (account: SavedAccount | User) =>
             </div>
           </DropdownMenuLabel>
           <DropdownMenuSeparator />
-          <div
-            v-if="scrobblingStatus.configured"
-            class="flex items-center gap-2 px-2 py-1.5 text-xs"
-            role="status"
-            :aria-label="scrobblingLabel"
-          >
-            <Activity
-              class="size-4 shrink-0 text-emerald-500"
-              aria-hidden="true"
-            />
-            <span class="min-w-0 truncate">
-              {{ $t("auth.scrobbling") }} · {{ providerSummary }}
-            </span>
-            <span class="sr-only">{{ scrobblingLabel }}</span>
-          </div>
-          <DropdownMenuSeparator v-if="scrobblingStatus.configured" />
           <DropdownMenuItem @click="handleProfile">
             <Pencil class="size-4" />
             {{ $t("auth.edit_profile") }}
@@ -431,110 +347,6 @@ const accountSwitcherClass = (account: SavedAccount | User) =>
       </DropdownMenu>
     </SidebarMenuItem>
   </SidebarMenu>
-
-  <Dialog v-model:open="switchAccountsOpen">
-    <DialogContent class="max-w-md">
-      <DialogHeader>
-        <DialogTitle>{{ $t("auth.switch_account") }}</DialogTitle>
-        <DialogDescription>{{
-          $t("auth.switch_account_description")
-        }}</DialogDescription>
-      </DialogHeader>
-      <div class="grid gap-2">
-        <button
-          v-for="item in availableAccounts"
-          :key="item.account?.token || item.user?.user_id"
-          type="button"
-          class="flex items-center gap-3 rounded-lg border border-l-4 p-3 text-left transition-colors hover:bg-accent"
-          :class="accountSwitcherClass(item.user || item.account!)"
-          @click="selectAccount(item)"
-        >
-          <Avatar
-            class="size-10 border"
-            :class="accountAccentClass(item.user || item.account!)"
-          >
-            <AvatarImage
-              v-if="item.user?.avatar_url || item.account?.avatarUrl"
-              :src="item.user?.avatar_url || item.account?.avatarUrl || ''"
-              :alt="accountName(item.user || item.account!)"
-            />
-            <AvatarFallback
-              class="text-primary-foreground"
-              :class="accountAccentClass(item.user || item.account!)"
-              >{{ accountInitial(item.user || item.account!) }}</AvatarFallback
-            >
-          </Avatar>
-          <span class="min-w-0 flex-1">
-            <span class="block truncate font-medium">{{
-              accountName(item.user || item.account!)
-            }}</span>
-            <span class="block truncate text-xs text-muted-foreground">
-              @{{ item.user?.username || item.account?.username }}
-            </span>
-          </span>
-          <span v-if="!item.account" class="text-xs text-muted-foreground">{{
-            $t("auth.login")
-          }}</span>
-          <span
-            v-if="item.account?.token === authManager.getToken()"
-            class="text-xs text-muted-foreground"
-            >{{ $t("auth.current") }}</span
-          >
-        </button>
-        <p
-          v-if="savedAccounts.length < 2"
-          class="rounded-lg bg-muted p-3 text-sm text-muted-foreground"
-        >
-          {{ $t("auth.add_account_hint") }}
-        </p>
-        <button
-          type="button"
-          class="rounded-lg border px-3 py-2 text-sm font-medium transition-colors hover:bg-accent"
-          @click="addAccount"
-        >
-          {{ $t("auth.add_account") }}
-        </button>
-      </div>
-    </DialogContent>
-  </Dialog>
-
-  <Dialog v-model:open="loginDialogOpen">
-    <DialogContent class="max-w-sm">
-      <DialogHeader>
-        <DialogTitle>{{ $t("auth.login") }}</DialogTitle>
-        <DialogDescription>{{ loginUser?.username }}</DialogDescription>
-      </DialogHeader>
-      <form class="grid gap-4" @submit.prevent="signInToAccount">
-        <div class="grid gap-2">
-          <label for="switch-account-password" class="text-sm font-medium">
-            {{ $t("auth.password") }}
-          </label>
-          <Input
-            id="switch-account-password"
-            v-model="loginPassword"
-            type="password"
-            autocomplete="current-password"
-            autofocus
-          />
-          <p v-if="loginError" class="text-sm text-destructive">
-            {{ loginError }}
-          </p>
-        </div>
-        <DialogFooter>
-          <Button
-            type="button"
-            variant="outline"
-            @click="loginDialogOpen = false"
-          >
-            {{ $t("cancel") }}
-          </Button>
-          <Button type="submit" :loading="loginLoading">
-            {{ $t("auth.login") }}
-          </Button>
-        </DialogFooter>
-      </form>
-    </DialogContent>
-  </Dialog>
 </template>
 
 <style scoped>
@@ -543,6 +355,7 @@ const accountSwitcherClass = (account: SavedAccount | User) =>
 }
 
 .scrobbling-avatar-glow {
+  filter: brightness(0.7);
   z-index: 1;
 }
 
@@ -551,6 +364,7 @@ const accountSwitcherClass = (account: SavedAccount | User) =>
 }
 
 .scrobbling-avatar-glow-base {
+  filter: brightness(0.7);
   opacity: 0.3;
   z-index: 1;
 }
